+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Make sure that `ServerMutateTicks.last_confirmed_tick` cannot go backwards.
+
 ## [0.42.2] - 2026-08-14
 
 ### Added